Iteration 2 of sound section
A redesign of the 'sound' section that provides a sound preview of each
available channel, so that people can choose the channel providing
the best sound.

The audio player is https://www.tympanus.net/Development/AudioPlayer.
Whichever audio player we pick, should be simple. The only options needed
are:

1. Play / pause button
2. Time stamps (time ellapsed / total duration)
3. Volume control
4. Track scrubbing (so that I can easily move back and forth)

The section also includes a 'there is no sound in any channel' option, and in this
case asks the speaker if s/he wants the video published anyway or not.

Signed-off-by: Belen Pena <belenbarrospena@gmail.com>