First design attempt for the video states
I have reduced the states listed at https://github.com/yoe/sreview#states to four:

1. Generating video
2. Ready for review
3. Processing
4. Published

Each of the above states in this prototype shows in its own page.

I have not included any indication of progress for the time being, because
I am keen on keeping it as simple as possible, and because I am not sure they
are worth the effort. If emails are sent to speakers after each significant
state change (from 'generating video' to 'ready for review' and from
'processing' to 'published'), I'd say that should be enough.

Signed-off-by: Belen Pena <belenbarrospena@gmail.com>