No data loss message and restore original video
Add some text reassuring people that nothing they can do in the form will result
in data loss, and an option to restore the video to its original state.

The latter option should only appear if changes have been made to the original
video. Otherwise, the sentence containing the link does not show.

I've made the restore option as inconspicuous as possible, and a 2-step process,
because this is a very destructive action. We don't want people to do this
by mistake.

Signed-off-by: Belen Pena <belenbarrospena@gmail.com>