Add a/v sync option
Add a section to report when audio and video are out of sync.

When the "audio is ahead" option is selected, the time difference value
should be negative.

When the "video is ahead" option is selected, the time difference
value should be positive.

This saves people from having to work it out by themselves, since it
is easy to make a mistake.

Signed-off-by: Belen Pena <belenbarrospena@gmail.com>