Changes to the "How is this video?" question
Remove the third option "Other brokenness" so that there are
only 2 answers to the question. The "Other brokenness" textarea
becomes part of the form displayed when you select the second
answer: "This video has problems".

This implements design recommendation no. 1 at
https://github.com/belenbarrospena/sreview_design/wiki/Design-feedback-sessions

Since I was there, I also reduced the size of the "Other brokenness"
textarea, which was massive.

Signed-off-by: Belen Pena <belenbarrospena@gmail.com>