Changes to thank you messages
The goal is to make it more visually obvious that everything has
gone according to plan, and to move the next step (the email)
closer to the top of the message.

After fiddling with it a bit, I removed the grey background, made the
heading green, and after trying a few icons from the set available,
I settled for the thumbs-up one because it conveys the success message
and looks good at the heading font size.

This implements design recommendation no.5 at
https://github.com/belenbarrospena/sreview_design/wiki/Design-feedback-sessions

Signed-off-by: Belen Pena <belenbarrospena@gmail.com>