Add radio button group to the 'Other brokenness' section
During user validation sessions, we came across the possibility of missuse
of the 'other brokenness' text area. Writing anything in there will
flag the video as a task for the video team. If people were to write
things there that are not related to video problems (e.g. a friendly
thank you to the video team for their hard work), that can create additional
workload for the video team.

We have therefore decided to remove the temptation of having the textarea
just there in front of you, and hide it behind a radio button group
that explicitily asks whether there are any additional problems with
the video that need to be reported.

This commit adds the radio button group.

Signed-off-by: Belen Pena <belenbarrospena@gmail.com>