First version of video table functionality
We have filter-as-you-type for talks and speakers, and dropdown
filters with set values for all other columns. We also have an
"Add talk" button.

There are two types of filters with set values:

1. Simple ones for dates and progress. They provide a list
of checkboxes to select from. We use the simple filter when
there are 5 or less values to choose from.

2. Complex ones for rooms, tracks and states. On top of the
list of checkboxes, they also provide a text input field
with filter-as-you-type functionality.

Signed-off-by: Belen Pena <belenbarrospena@gmail.com>